add c; remove file after writing

+++ src/executor.ts
@@ -1,11 +1,18 @@
 import { platform, tmpdir } from "os";
-import { writeFileSync } from "fs";
+import { writeFileSync, unlinkSync } from "fs";
 import { exec } from "child_process";
 import Language from "./languages";
+import { sync as commandExists } from "command-exists";
 const uniqueFilename = require("unique-filename");
 
+interface ISecurity {
+  enabled?: boolean;
+  uselxc?: boolean;
+  timeout?: number;
+}
+
 interface IOptions {
-  security: boolean;
+  security: ISecurity;
 }
 
 /**
@@ -26,8 +33,14 @@ export default async function execute(
     throw Error("Your OS is not supported yet.");
   }
 
+  // Check if Command Exists
+  if (!commandExists(language)) {
+    throw new Error("This language is not installed on the machine");
+  }
+
   // Write File to temp folder
-  const filepath: string = uniqueFilename(tmpdir());
+  var filepath: string = uniqueFilename(tmpdir());
+  if (language == Language.C) filepath += ".c";
   writeFileSync(filepath, input, { encoding: "utf-8" });
 
   // Execute code
@@ -35,6 +48,9 @@ export default async function execute(
     exec(
       `sh ${__dirname}/../runners/${os}/${language}.sh ${filepath}`,
       (err, stdout, stderr) => {
+        // Delete created file
+        unlinkSync(filepath);
+
         if (stderr) return reject(stderr);
         resolve(stdout);
       }
